Remove ClientBuilder trait
Previously, we used our own ClientBuilder trait as a wrapper around
trussed’s own ClientBuilder struct so that Apps implementors don’t need
to pass around the service and syscall.  But this limits the way the
runner can access the service, and is incompatible with recent changes
to the Nitrokey 3 firmware runner:
        https://github.com/Nitrokey/nitrokey-3-firmware/pull/504

Therefore, this patch removes the ClientBuilder trait.  Apps
implementers directly receive the Service and Syscall and can use
trussed::client::ClientBuilder to construct the clients, like they would
normally.