Replace FIDO example with a dummy app
Maintaining a working setup for the FIDO and admin app is a lot of work
and is redundant to the usbip runner provided as part of the Nitrokey 3
firmware.  Therefore, this patch replaces the fido example with a much
simpler example that just sets up a dummy client responding to a vendor
command that requests random data.

This is sufficient for basic testing and getting started.  For more
advanced use cases, the Nitrokey 3 runner should be used.

Fixes: https://github.com/trussed-dev/pc-usbip-runner/issues/31